+++ userdata/rhel.sh
@@ -127,6 +127,10 @@ else
   yum install -y postgresql mysql redis;
 fi
 
+# RDS Certificate Authority
+mkdir /root/.postgresql;
+wget -O /root/.postgresql/root.crt https://truststore.pki.rds.amazonaws.com/global/global-bundle.pem;
+
 #########################
 ##### Node.js v24.x #####
 #########################